migrate package to src/ structure
